[refactor] ReviewSortCondition -> SortCondition 으로 변경 후 패키지 global 로 이동
