+++ src/main/java/com/inq/wishhair/wesharewishhair/global/utils/PageableGenerator.java
@@ -4,7 +4,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 
-import static com.inq.wishhair.wesharewishhair.review.utils.ReviewSortCondition.DATE;
+import static com.inq.wishhair.wesharewishhair.global.utils.SortCondition.DATE;
 
 public interface PageableGenerator {
     static Pageable getDefaultPageable() {
