+++ src/main/java/com/inq/wishhair/wesharewishhair/review/controller/ReviewSearchController.java
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static com.inq.wishhair.wesharewishhair.review.utils.ReviewSortCondition.*;
+import static com.inq.wishhair.wesharewishhair.global.utils.SortCondition.*;
 
 @RestController
 @RequiredArgsConstructor
