+++ src/test/java/com/inq/wishhair/wesharewishhair/global/utils/DefaultPageableUtils.java
@@ -4,8 +4,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 
-import static com.inq.wishhair.wesharewishhair.review.utils.ReviewSortCondition.DATE;
-import static com.inq.wishhair.wesharewishhair.review.utils.ReviewSortCondition.LIKES;
+import static com.inq.wishhair.wesharewishhair.global.utils.SortCondition.DATE;
+import static com.inq.wishhair.wesharewishhair.global.utils.SortCondition.LIKES;
 
 public abstract class DefaultPageableUtils {
     public static Pageable getLikeDescPageable(int size) {
